do not lose data in case of already migrated users

+++ proso_user/models.py
@@ -225,8 +225,10 @@ def migrate_google_openid_user(user):
             new_user = new_social.user
             new_social.user = old_social.user
             new_social.save()
-            new_user.delete()
-            old_social.delete()
+            # in case of already migrated users do not lose data
+            if new_user.id != old_social.user.id:
+                new_user.delete()
+                old_social.delete()
             LOGGER.info('Migrating user "{}" from Google OpenID to OAauth2'.format(user.email))
             return old_social.user
         except UserSocialAuth.DoesNotExist:
